+++ matchbox-frontend/src/app/app.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { FhirConfigService } from './fhirConfig.service';
 import { TranslateService } from '@ngx-translate/core';
 import packageJson from '../../package.json';
+import {HashUrlRedirectionService} from "./util/hash-url-redirection-service";
 
 @Component({
   selector: 'app-root',
@@ -11,17 +12,24 @@ import packageJson from '../../package.json';
 export class AppComponent {
   public version: string = packageJson.version;
 
-  constructor(translateService: TranslateService, fhirConfigService: FhirConfigService) {
+  constructor(readonly translateService: TranslateService,
+              readonly fhirConfigService: FhirConfigService,
+              readonly hashUrlRedirectionService: HashUrlRedirectionService) {
+    // Redirect any old URL with hash to the new URL
+    if (hashUrlRedirectionService.isHashUrl()) {
+      hashUrlRedirectionService.redirectHashUrl();
+    }
+
     translateService.setDefaultLang('de');
     translateService.use(translateService.getBrowserLang());
 
     let base = location.origin;
     if (base === 'http://localhost:4200') {
-      console.log('note: using local dev mag system for' + location.origin);
+      console.log('note: using local dev mag system for ' + location.origin);
       // You can also use /proxy/testahdisch
       fhirConfigService.changeFhirMicroService('http://localhost:4200/proxy/localhost/matchboxv3/fhir');
     } else {
-      let url: string = base + location.pathname + 'fhir';
+      const url: string = (window as any).MATCHBOX_BASE_PATH + '/fhir';
       fhirConfigService.changeFhirMicroService(url);
       console.log('fhir endpoint ' + url);
     }