+++ matchbox-server/src/main/java/ca/uhn/fhir/jpa/starter/Application.java
@@ -1,5 +1,6 @@
 package ca.uhn.fhir.jpa.starter;
 
+import ch.ahdis.matchbox.config.MatchboxStaticResourceConfig;
 import ch.ahdis.matchbox.spring.MatchboxEventListener;
 import ch.ahdis.matchbox.terminology.RegistryWs;
 import ch.ahdis.matchbox.validation.gazelle.GazelleValidationWs;
@@ -31,7 +32,8 @@ import ch.ahdis.matchbox.config.MatchboxJpaConfig;
 	FhirServerConfigR4.class,
 	MatchboxEventListener.class,
 	GazelleValidationWs.class,
-  RegistryWs.class})
+  RegistryWs.class,
+  MatchboxStaticResourceConfig.class})
 public class Application extends SpringBootServletInitializer {
 
   public static void main(String[] args) {